requirements.txt
pyserial==3.5
pillow==10.4.0
pyqt6==6.7.0
pynput>=1.7.7

